requirements.txt
gevent==1.3.6
django==2.1.1
leancloud==2.1.8
django-simple-captcha==0.5.9
pymysql==0.9.2